Добавлены файлы БД и ПЗ Изменена работа с данными (все данные временно размещаются внутри самого приложения, потому что имеются проблемы с локальным сервером баз данных) Изменён дизайн (сделан на скорую руку и будет дальше дорабатываться)

+++ GostControl/AppViewModels/MainViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using GostControl.AppModels;
 using GostControl.AppRepositories;
+using GostControl.AppWindows;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Windows.Input;
@@ -35,6 +36,7 @@ namespace GostControl.AppViewModels
         public ICommand LoadClientsCommand { get; }
         public ICommand AddClientCommand { get; }
         public ICommand DeleteClientCommand { get; }
+        public ICommand RefreshCommand { get; }
 
         public MainViewModel()
         {
@@ -44,6 +46,7 @@ namespace GostControl.AppViewModels
             LoadClientsCommand = new RelayCommand(LoadClients);
             AddClientCommand = new RelayCommand(AddClient);
             DeleteClientCommand = new RelayCommand(DeleteClient, CanDeleteClient);
+            RefreshCommand = new RelayCommand(RefreshData);
 
             LoadClients(null);
         }
@@ -60,16 +63,48 @@ namespace GostControl.AppViewModels
 
         private void AddClient(object parameter)
         {
-            LoadClients(null);
+            // Создаем нового клиента
+            var newClient = new Client
+            {
+                ClientID = 0,
+                LastName = "Новый",
+                FirstName = "Клиент",
+                MiddleName = "",
+                PhoneNumber = "+7",
+                Email = "email@example.com",
+                DateOfBirth = System.DateTime.Now.AddYears(-30),
+                RegistrationDate = System.DateTime.Now
+            };
+
+            var addWindow = new AddEditClientWindow(newClient, true);
+            if (addWindow.ShowDialog() == true)
+            {
+                _clientRepository.AddClient(newClient);
+                LoadClients(null);
+            }
         }
 
         private void DeleteClient(object parameter)
         {
             if (SelectedClient != null)
             {
-                _clientRepository.DeleteClient(SelectedClient.ClientID);
-                LoadClients(null);
+                var result = System.Windows.MessageBox.Show(
+                    $"Удалить клиента {SelectedClient.FullName}?",
+                    "Подтверждение удаления",
+                    System.Windows.MessageBoxButton.YesNo,
+                    System.Windows.MessageBoxImage.Question);
+
+                if (result == System.Windows.MessageBoxResult.Yes)
+                {
+                    _clientRepository.DeleteClient(SelectedClient.ClientID);
+                    LoadClients(null);
+                }
             }
+        }
+
+        private void RefreshData(object parameter)
+        {
+            LoadClients(null);
         }
 
         private bool CanDeleteClient(object parameter)
